gradient descent

+++ chess4j-java/src/main/java/com/jamesswafford/chess4j/tuner/Hypothesis.java
@@ -11,7 +11,7 @@ public class Hypothesis {
         return hypothesis(score);
     }
 
-    public static double hypothesis(int score) {
+    public static double hypothesis(double score) {
         // This is the traditional approach: 1 / (1 + e ^ -z)
         //return 1.0 / (1 + Math.exp(-score));
